individual_apps/Smart-Course-Search/requirements.txt
gradio>=3.50.0
langchain>=0.1.0
langchain-community>=0.0.10
python-dotenv>=0.21.0
pyyaml>=6.0
together>=0.2.5
pinecone-client>=2.2.4
markdown>=3.4.0
python-multipart>=0.0.6
typing-extensions>=4.5.0
typing-inspect>=0.9.0
aiohttp>=3.8.0
async-timeout>=4.0.0
requests>=2.31.0
urllib3>=1.26.0
PyYAML>=6.0.1
tqdm>=4.66.0
numpy>=1.24.0
pandas>=2.0.0
